price_fetcher/pyproject.toml
[tool.poetry]
name = "price-fetcher"
version = "0.1.0"
description = "to fetch the price from multiple sources, currently fetches close price only."
authors = ["archit pandita"]
readme = "README.md"
packages = [{include = "price_fetcher"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pydantic = "^2.9.2"
pyyaml = "^6.0.2"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
